refactor: address task details review feedback
